Link xinerama_pic. closes: #189476
As a library, libxosd needs to be linkted with "Position Independent
Code"-version of libxinerama, since there is no shared library of
libxinerama, yet. Below you will find the patch, I do have to apply for
Debian, but I don't like it, because I don't know if every other
distribution has a lib*_pic.a library.


git-svn-id: svn://svn.code.sf.net/p/libxosd/code@315 3a7ed7e4-6e13-0410-ba71-8ed462f9d0c2