Fix XFreeFontSet() after invalid font. (closes: #210680)
[BUG]
This bug was reportes via Debian as #210680 making xmms crash.
The bug manifests as following: During initialization of the xmms plugin, the
default font is chosen, so the xosd context is successfully created. In
ally_config(), an invalid font is loaded from xmms's config file and
xosd->fontset gets set to NULL, so xosd doesn't seem to work. On disableing the
plugin, NULL is freed and xmms crashes.


git-svn-id: svn://svn.code.sf.net/p/libxosd/code@377 3a7ed7e4-6e13-0410-ba71-8ed462f9d0c2